dll detach, suspending other threads
other improvements

+++ src/lib.rs
@@ -1,3 +1,5 @@
+use winapi::um::winnt::PAGE_EXECUTE_READWRITE;
+use windows::Win32::System::LibraryLoader::FreeLibraryAndExitThread;
 use windows::Win32::{
     Foundation::*, System::SystemServices::*, UI::WindowsAndMessaging::MessageBoxA,
 };
@@ -5,28 +7,31 @@ use windows::core::*;
 
 use winapi::{
     shared::minwindef::{DWORD, HINSTANCE, LPVOID},
-    um::consoleapi::AllocConsole,
+    um::{consoleapi::AllocConsole, memoryapi::{VirtualProtect}},
+    
 };
 
-use std::{slice, thread::{self}, time};
-
+use std::sync::atomic::{AtomicIsize, AtomicUsize, Ordering::SeqCst};
+use std::{slice, thread::{self}, time, ffi::c_void};
 // modules
 #[macro_use]
 mod log; use log::*;
 mod val; use val::*;
 mod mem; use mem::*;
 
-
+static MY_HANDLE: AtomicUsize = AtomicUsize::new(0); // global var to store this DLL's handle
 // entry. "system" defines the calling convention
 #[unsafe(no_mangle)]
 unsafe extern "system" fn DllMain(handle: HINSTANCE, call_reason: DWORD, _: LPVOID) -> bool {
     match call_reason {
         DLL_PROCESS_ATTACH => unsafe {
+            MY_HANDLE.store(handle as usize, SeqCst);
             MessageBoxA(None, s!("attached"), s!("spearman.dll"), Default::default());
             // do stuff in other thread so we don't freeze process (return from DllMain immediately)
             thread::spawn(|| begin());
         }
         DLL_PROCESS_DETACH => unsafe{
+            info!("Detached.");
             unregister_dll_hook()
         },
         _ => (),
@@ -55,27 +60,39 @@ pub unsafe fn on_dll() { unsafe{
         }
     };
 
-    info!("Successfully located {}.", TARGET);
+    good!("Successfully located {}.", TARGET);
     info!("-----------------");
     info!("Base address: 0x{:x}", modu.info.lpBaseOfDll as u64);
     info!("Size of image: {} B", modu.info.SizeOfImage);
     info!("-----------------\n");
 
     info!("Signature: {:02X?}", PATTERN);
-    info!("Scanning...");
 
+    info!("Suspending other threads...");
+    let handles = sus_threads();
+
+    info!("Scanning...");
     let addr = scan(&PATTERN, modu.info.lpBaseOfDll as *mut u8, modu.info.SizeOfImage as usize);
 
-    info!("Pattern matched at address 0x{:x}...", addr as u32);
+    good!("Pattern matched at address 0x{:x}...", addr as u32);
 
-    info!("Patching signature check...");
-    addr.write_bytes(NOP, PATTERN.len());
-    info!("Replaced {} bytes with {}.", PATTERN.len(), NOP);
+    
+    info!("Getting access PAGE_EXECUTE_READWRITE...");
+    let mut old: u32 = 0;
+    if VirtualProtect(addr as _, PATTERN.len(), PAGE_EXECUTE_READWRITE, &mut old) == 0 { die!("Failed to get PAGE_EXECUTE_READWRITE access using VirtualProtect.") }
+        addr.write_bytes(NOP, PATTERN.len());
+        info!("Replaced {} bytes with {}.", PATTERN.len(), NOP);
+        info!("Reverting protection...");
+    if VirtualProtect(addr as _, PATTERN.len(), old, &mut old) == 0 { error!("Failed to revert protection level. It will stay as PAGE_EXECUTE_READWRITE.") };
 
-    info!("Patch complete; detaching!");
+    info!("Resuming other threads...");
+    res_threads(handles);
 
     // detach dll
-    return
+    good!("Patch complete!");
+    info!("Press enter to detach...");
+    let _ = std::io::stdin().read_line(&mut String::new());
+    FreeLibraryAndExitThread(HMODULE(MY_HANDLE.load(SeqCst) as *mut c_void), 0);
 }}
 
 /// brute force scanning method. *mut u8 is a byte pointer
